refactor(evals): shared den-web helpers for cloud-connection flows
The three cloud-connection flows each carried identical copies of
denApiFetch, API/browser sign-in, and the Extensions>Connections /
Your Connections navigation. They now import evals/flows/lib/den-web.mjs
(the runner only loads top-level *.flow.mjs, so lib/ is inert). Also:
cloud-oauth now deletes leftover fraimz-mcp-* rows at setup so reruns
don't accumulate connections. Net -71 lines, behavior unchanged — all
three flows re-ran PASSED.
