+++ evals/flows/mcp-connections-cloud-oauth.flow.mjs
@@ -40,30 +40,13 @@
  *   accommodation, not a product behavior change).
  */
 
-const DEN_API_URL = (process.env.OPENWORK_EVAL_DEN_API_URL ?? "").trim().replace(/\/+$/, "");
-const DEN_WEB_URL = (process.env.OPENWORK_EVAL_DEN_WEB_URL ?? "").trim().replace(/\/+$/, "");
+import { denApiFetch, denApiUrl, denWebUrl, openAdminConnections, signInApi, signInViaBrowser } from "./lib/den-web.mjs";
+
 const DEMO_EMAIL = process.env.OPENWORK_EVAL_DEMO_EMAIL?.trim() || "alex@acme.test";
 const DEMO_PASSWORD = process.env.OPENWORK_EVAL_DEMO_PASSWORD?.trim() || "OpenWorkDemo123!";
 const MOCK_SERVER_URL = (process.env.MOCK_OAUTH_MCP_URL ?? "http://127.0.0.1:3978").trim().replace(/\/+$/, "");
 const CONNECTION_NAME = `fraimz-mcp-${Date.now()}`;
 const ECHO_TEXT = "search and execute in the cloud proof";
-
-async function denApiFetch(path, options = {}) {
-  const response = await fetch(`${DEN_API_URL}${path}`, {
-    ...options,
-    // Better Auth rejects auth requests with no Origin header (CSRF
-    // protection); a real browser always sends one, Node's fetch doesn't.
-    headers: { "content-type": "application/json", origin: DEN_WEB_URL, ...(options.headers ?? {}) },
-  });
-  const text = await response.text();
-  let body;
-  try {
-    body = JSON.parse(text);
-  } catch {
-    body = text;
-  }
-  return { response, body };
-}
 
 export default {
   id: "mcp-connections-cloud-oauth",
@@ -77,9 +60,25 @@ export default {
       run: async (ctx) => {
         const health = await fetch(`${MOCK_SERVER_URL}/health`).catch(() => null);
         ctx.assert(Boolean(health?.ok), `Mock OAuth+MCP server not reachable at ${MOCK_SERVER_URL}.`);
+        const adminSession = await signInApi(DEMO_EMAIL, DEMO_PASSWORD);
+        ctx.assert(Boolean(adminSession), `Den API sign-in failed for ${DEMO_EMAIL}.`);
+        const existing = await denApiFetch("/v1/mcp-connections?scope=manageable", {
+          headers: { authorization: `Bearer ${adminSession}` },
+        });
+        ctx.assert(existing.response.ok, `Listing manageable connections failed: ${existing.response.status}`);
+        for (const connection of existing.body.connections ?? []) {
+          if (connection.name.startsWith("fraimz-mcp-")) {
+            const removed = await denApiFetch(`/v1/mcp-connections/${connection.id}`, {
+              method: "DELETE",
+              headers: { authorization: `Bearer ${adminSession}` },
+            });
+            ctx.assert(removed.response.ok, `Cleanup delete failed for leftover ${connection.id}.`);
+          }
+        }
+        const webUrl = denWebUrl();
         const currentUrl = await ctx.eval("window.location.href");
-        if (!currentUrl.includes(new URL(DEN_WEB_URL).host)) {
-          await ctx.eval(`(() => { window.location.href = ${JSON.stringify(DEN_WEB_URL)}; return true; })()`);
+        if (!currentUrl.includes(new URL(webUrl).host)) {
+          await ctx.eval(`(() => { window.location.href = ${JSON.stringify(webUrl)}; return true; })()`);
         }
         await ctx.waitFor("document.readyState === 'complete'", { timeoutMs: 30_000, label: "den-web page loaded" });
       },
@@ -97,51 +96,13 @@ export default {
           ctx.log("Already signed in as an admin; reusing session.");
           return;
         }
-        await ctx.eval(`fetch('/api/auth/sign-out', { method: 'POST', headers: { 'content-type': 'application/json' }, body: '{}' }).then(() => true).catch(() => true)`, { awaitPromise: true });
-        await ctx.eval(`(() => { window.location.href = ${JSON.stringify(DEN_WEB_URL)}; return true; })()`);
-        await ctx.waitFor("document.readyState === 'complete'", { timeoutMs: 30_000 });
-        await ctx.waitFor(
-          "Boolean(document.querySelector('input[type=\"email\"]')) && Boolean(document.querySelector('input[type=\"password\"]'))",
-          { timeoutMs: 30_000, label: "email + password fields" },
-        );
-        await ctx.fill('input[type="email"]', DEMO_EMAIL);
-        await ctx.fill("input[type=\"password\"]", DEMO_PASSWORD);
-        const submitted = await ctx.eval(`(() => {
-          const button = document.querySelector('button[type="submit"]');
-          if (!button) return false;
-          button.click();
-          return true;
-        })()`);
-        ctx.assert(submitted, "No submit button found on the sign-in card.");
-        await ctx.waitForText("Dashboard", { timeoutMs: 30_000 });
+        await signInViaBrowser(ctx, DEMO_EMAIL, DEMO_PASSWORD);
       },
     },
     {
       name: "Open Settings -> MCP Connections",
       run: async (ctx) => {
-        // Retry the click: on a fresh page load (especially over higher cloud
-        // latency), the very first click can land before Next.js has finished
-        // hydrating and attaching the link's handler. The Connections link
-        // lives inside the collapsible Extensions nav group, so expand that
-        // first when the link isn't in the DOM yet.
-        await ctx.waitFor(
-          `(() => {
-            if (window.location.pathname.includes('mcp-connections')) return true;
-            const link = [...document.querySelectorAll('nav a')].find((a) => a.getAttribute('href')?.includes('mcp-connections'));
-            if (link) {
-              link.click();
-              return false;
-            }
-            const group = [...document.querySelectorAll('nav a, nav button')].find((el) => (el.textContent ?? '').trim().startsWith('Extensions'));
-            group?.click();
-            return false;
-          })()`,
-          { timeoutMs: 30_000, label: "MCP Connections nav link clicked" },
-        );
-        await ctx.waitFor("window.location.pathname.includes('mcp-connections')", {
-          timeoutMs: 20_000,
-          label: "MCP Connections route",
-        });
+        await openAdminConnections(ctx);
         await ctx.prove("The Connections screen renders in Den", {
           assert: async () => {
             await ctx.expectText("Add a custom MCP server");
@@ -276,7 +237,7 @@ export default {
             const mcpToken = mint.body.token;
 
             async function mcpAgentCall(method, params) {
-              const response = await fetch(`${DEN_API_URL}/mcp/agent`, {
+              const response = await fetch(`${denApiUrl()}/mcp/agent`, {
                 method: "POST",
                 headers: {
                   "content-type": "application/json",
